+++ src/game/Server/Packets/Battleground.cpp
@@ -54,5 +54,5 @@ void WorldPackets::Battleground::BattlefieldJoin::ReadFromWorldPacket(WorldPacke
 
 void WorldPackets::Battleground::GroupJoinedBattleground::AppendBodyTo(ByteBuffer& buffer) const
 {
-    buffer << uint32(result);
+    buffer << result;
 }
