+++ src/game/Server/Packets/Spell.cpp
@@ -37,7 +37,7 @@ void WorldPackets::Spell::CancelChanneling::ReadFromWorldPacket(WorldPacket& rec
 
 void WorldPackets::Spell::CastResultSimpleFailure::AppendBodyTo(ByteBuffer& buffer) const
 {
-    buffer << uint32(spellId);
+    buffer << spellId;
     buffer << uint8(2); // status = fail
     buffer << uint8(reason);
 }
